tests: reorganise tests by tier
