+++ src/poly/tests/metrics_test.py
@@ -3,22 +3,10 @@
 Copyright PolyAI Limited
 """
 
-import json
-import os
-import shutil
-import tempfile
 import unittest
-from copy import deepcopy
 from unittest.mock import MagicMock, patch
 
 from poly.cli_commands.metrics import VALID_METRIC_TYPES, MetricsCommand, _parse_bool_flag
-from poly.handlers.interface import AgentStudioInterface
-from poly.project import AgentStudioProject
-
-DIR = os.path.dirname(os.path.abspath(__file__))
-EMPTY_PROJECT_DIR = os.path.join(DIR, "test_projects", "test_empty_project")
-EMPTY_PROJECT_DATA_LOC = os.path.join(EMPTY_PROJECT_DIR, "empty_project.json")
-EMPTY_PROJECT_DATA = json.loads(open(EMPTY_PROJECT_DATA_LOC, encoding="utf-8").read())
 
 
 class ParseBoolFlagTest(unittest.TestCase):
@@ -810,194 +798,6 @@ class PrintMetricsTest(unittest.TestCase):
         summary = mock_console.print.call_args_list[1][0][0]
         self.assertIn("1 active", summary)
         self.assertIn("1 inactive", summary)
-
-
-class ProjectCreateCustomMetricTest(unittest.TestCase):
-    """Tests for AgentStudioProject.create_custom_metric validation and orchestration."""
-
-    def setUp(self):
-        self.temp_dir = tempfile.mkdtemp()
-        self.project = AgentStudioProject.from_dict(deepcopy(EMPTY_PROJECT_DATA), self.temp_dir)
-
-    def tearDown(self):
-        shutil.rmtree(self.temp_dir, ignore_errors=True)
-
-    @patch("poly.project.AgentStudioInterface.set_custom_metric_api_flag")
-    @patch("poly.project.AgentStudioInterface.create_custom_metric")
-    def test_api_flag_triggers_follow_up_update(self, mock_create, mock_set_api):
-        """When api=True, a follow-up call sets the api flag after create."""
-        mock_create.return_value = {"name": "SCORE", "type": "int"}
-        mock_set_api.return_value = {"name": "SCORE", "type": "int", "api": True}
-
-        result = self.project.create_custom_metric({"name": "SCORE", "type": "int", "api": True})
-
-        mock_create.assert_called_once()
-        mock_set_api.assert_called_once_with(
-            self.project.region, self.project.account_id, self.project.project_id, "SCORE", True
-        )
-        self.assertTrue(result["api"])
-
-    @patch("poly.project.AgentStudioInterface.set_custom_metric_api_flag")
-    @patch("poly.project.AgentStudioInterface.create_custom_metric")
-    def test_no_api_flag_skips_follow_up(self, mock_create, mock_set_api):
-        """When api is not set, no follow-up call is issued."""
-        mock_create.return_value = {"name": "SCORE", "type": "int"}
-
-        self.project.create_custom_metric({"name": "SCORE", "type": "int"})
-
-        mock_create.assert_called_once()
-        mock_set_api.assert_not_called()
-
-    def test_expected_values_rejected_for_non_string(self):
-        """Raises ValueError when expected_values is set on a non-string metric."""
-        with self.assertRaises(ValueError) as ctx:
-            self.project.create_custom_metric(
-                {"name": "SCORE", "type": "int", "expected_values": ["a", "b"]},
-            )
-
-        self.assertIn("only valid for string", str(ctx.exception))
-
-    @patch("poly.project.AgentStudioInterface.create_custom_metric")
-    def test_expected_values_allowed_for_string(self, mock_create):
-        """Does not raise when expected_values is set on a string metric."""
-        mock_create.return_value = {"name": "STATUS", "type": "string"}
-
-        self.project.create_custom_metric(
-            {"name": "STATUS", "type": "string", "expected_values": ["open", "closed"]},
-        )
-
-        mock_create.assert_called_once()
-
-
-class ProjectUpdateCustomMetricTest(unittest.TestCase):
-    """Tests for AgentStudioProject.update_custom_metric validation and orchestration."""
-
-    def setUp(self):
-        self.temp_dir = tempfile.mkdtemp()
-        self.project = AgentStudioProject.from_dict(deepcopy(EMPTY_PROJECT_DATA), self.temp_dir)
-
-    def tearDown(self):
-        shutil.rmtree(self.temp_dir, ignore_errors=True)
-
-    @patch("poly.project.AgentStudioInterface.update_custom_metric")
-    @patch("poly.project.AgentStudioInterface.get_custom_metrics")
-    def test_expected_values_rejected_for_non_string(self, mock_get, mock_update):
-        """Raises ValueError when expected_values targets a non-string metric."""
-        mock_get.return_value = [{"name": "SCORE", "type": "int"}]
-
-        with self.assertRaises(ValueError) as ctx:
-            self.project.update_custom_metric("SCORE", {"expected_values": ["a", "b"]})
-
-        self.assertIn("only valid for string", str(ctx.exception))
-        mock_update.assert_not_called()
-
-    @patch("poly.project.AgentStudioInterface.update_custom_metric")
-    @patch("poly.project.AgentStudioInterface.get_custom_metrics")
-    def test_expected_values_allowed_for_string(self, mock_get, mock_update):
-        """Does not raise when expected_values targets a string metric."""
-        mock_get.return_value = [{"name": "STATUS", "type": "string"}]
-        mock_update.return_value = {"name": "STATUS"}
-
-        self.project.update_custom_metric("STATUS", {"expected_values": ["open"]})
-
-        mock_update.assert_called_once()
-
-    @patch("poly.project.AgentStudioInterface.update_custom_metric")
-    def test_no_expected_values_skips_type_check(self, mock_update):
-        """When expected_values is not in data, no type lookup is made."""
-        mock_update.return_value = {"name": "SCORE"}
-
-        self.project.update_custom_metric("SCORE", {"description": "new desc"})
-
-        mock_update.assert_called_once()
-
-
-class ImportMetricsFromFileInterfaceTest(unittest.TestCase):
-    """Tests for AgentStudioInterface.import_metrics_from_file orchestration."""
-
-    @patch("poly.handlers.interface.PlatformAPIHandler.preview_metrics_import")
-    def test_dry_run_returns_preview(self, mock_preview):
-        """In dry-run mode, returns preview without importing."""
-        mock_preview.return_value = {
-            "would_create": ["SCORE"],
-            "would_skip": [],
-            "remote_only": [],
-        }
-
-        result = AgentStudioInterface.import_metrics_from_file(
-            "us", "acc1", "proj1", "SCORE:\n  type: int\n", {"SCORE"}, dry_run=True
-        )
-
-        self.assertTrue(result["dry_run"])
-        self.assertEqual(result["would_create"], ["SCORE"])
-
-    @patch("poly.handlers.interface.PlatformAPIHandler.import_custom_metrics")
-    @patch("poly.handlers.interface.PlatformAPIHandler.preview_metrics_import")
-    def test_import_returns_result_with_remote_only(self, mock_preview, mock_import):
-        """Full import merges remote_only from preview into the result."""
-        mock_preview.return_value = {
-            "would_create": ["SCORE"],
-            "would_skip": [],
-            "remote_only": ["OLD_METRIC"],
-        }
-        mock_import.return_value = {
-            "metadata": {"created": ["SCORE"], "ignored": []},
-        }
-
-        result = AgentStudioInterface.import_metrics_from_file(
-            "us", "acc1", "proj1", "SCORE:\n  type: int\n", {"SCORE"}, dry_run=False
-        )
-
-        self.assertEqual(result["remote_only"], ["OLD_METRIC"])
-        self.assertEqual(result["metadata"]["created"], ["SCORE"])
-        mock_import.assert_called_once()
-
-
-class ProjectImportMetricsFromFileTest(unittest.TestCase):
-    """Tests for AgentStudioProject.import_metrics_from_file file reading and delegation."""
-
-    def setUp(self):
-        self.temp_dir = tempfile.mkdtemp()
-        self.project = AgentStudioProject.from_dict(deepcopy(EMPTY_PROJECT_DATA), self.temp_dir)
-
-    def tearDown(self):
-        shutil.rmtree(self.temp_dir, ignore_errors=True)
-
-    def test_file_not_found_raises(self):
-        """Raises FileNotFoundError for a missing file."""
-        with self.assertRaises(FileNotFoundError):
-            self.project.import_metrics_from_file("/nonexistent/metrics.yaml")
-
-    def test_invalid_yaml_raises(self):
-        """Raises ValueError for unparseable YAML."""
-        bad_file = os.path.join(self.temp_dir, "bad.yaml")
-        with open(bad_file, "w") as f:
-            f.write("{{invalid")
-
-        with self.assertRaises(ValueError) as ctx:
-            self.project.import_metrics_from_file(bad_file)
-
-        self.assertIn("Invalid YAML", str(ctx.exception))
-
-    @patch("poly.project.AgentStudioInterface.import_metrics_from_file")
-    def test_delegates_parsed_content_to_interface(self, mock_import):
-        """Reads and parses the file, then delegates to the interface layer."""
-        metrics_file = os.path.join(self.temp_dir, "metrics.yaml")
-        with open(metrics_file, "w") as f:
-            f.write("SCORE:\n  type: int\n")
-        mock_import.return_value = {"metadata": {"created": ["SCORE"], "ignored": []}}
-
-        result = self.project.import_metrics_from_file(metrics_file, dry_run=True)
-
-        mock_import.assert_called_once_with(
-            self.project.region,
-            self.project.account_id,
-            self.project.project_id,
-            "SCORE:\n  type: int\n",
-            {"SCORE"},
-            True,
-        )
-        self.assertEqual(result["metadata"]["created"], ["SCORE"])
 
 
 class ValidMetricTypesTest(unittest.TestCase):
